requirements.txt
fastapi[standard]
uvicorn
sqlmodel
psycopg2-binary
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
ruff
black
pytest
httpx
